fix error tracing issue
Fix error tracing issue causing indels misplacement
