+++ PostProcess/post_process.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-set -e
-
 #$1 is CFD targets file
 #$2 is genecode annotation
 #$3 is empirical data for the guide
@@ -42,10 +40,7 @@ echo 'Sorting final annotation results using original NR to correspond with orig
 # LC_ALL=C sort -T $dir -k4,4rg $1.found.bed -o $1.found.bed
 LC_ALL=C sort -T $dir -k4,4n $1.found.bed -o $1.found.bed
 echo 'Starting integration with empirical data (this may take a while)'
-"$starting_dir"./resultIntegrator.py $1 $3 $1.found.bed $4 $6/ true $5 $7 $9 ${10} ${11} || {
-    echo "CRISPRme ERROR: result integration failed (script: ${0} line $((LINENO-1)))" >&2
-	exit 1
-}
+"$starting_dir"./resultIntegrator.py $1 $3 $1.found.bed $4 $6/ true $5 $7 $9 ${10} ${11}
 echo 'Removing unnecessary files'
 rm -f $1.bed $1.found.bed $1.redirectFile.out $1.temp.bed
 # sed -i 1i"#Bulge_type\tcrRNA\tDNA\tReference\tChromosome\tPosition\tCluster_Position\tDirection\tMismatches\tBulge_Size\tTotal\tPAM_gen\tVar_uniq\tSamples\tAnnotation_Type\tReal_Guide\trsID\tAF\tSNP\t#Seq_in_cluster\tCFD\tCFD_ref\tHighest_CFD_Risk_Score\tHighest_CFD_Absolute_Risk_Score\tMMBLG_#Bulge_type\tMMBLG_crRNA\tMMBLG_DNA\tMMBLG_Reference\tMMBLG_Chromosome\tMMBLG_Position\tMMBLG_Cluster_Position\tMMBLG_Direction\tMMBLG_Mismatches\tMMBLG_Bulge_Size\tMMBLG_Total\tMMBLG_PAM_gen\tMMBLG_Var_uniq\tMMBLG_Samples\tMMBLG_Annotation_Type\tMMBLG_Real_Guide\tMMBLG_rsID\tMMBLG_AF\tMMBLG_SNP\tMMBLG_#Seq_in_cluster\tMMBLG_CFD\tMMBLG_CFD_ref\tMMBLG_CFD_Risk_Score\tMMBLG_CFD_Absolute_Risk_Score" "$1"
