+++ pages/main_page.py
@@ -394,9 +394,7 @@ def change_url(
     gencode_name = "gencode.protein_coding.bed"
     annotation_name = ".dummy.bed"  # to proceed without annotation file
     if "EN" in annotation_var:
-        # annotation_name = "encode+gencode.hg38.bed"
-        ##use annotation wiht dhs
-        annotation_name = "dhs+encode+gencode.hg38.bed"
+        annotation_name = "dhs+encode+gencode.hg38.bed"  # use dhs annotation file
         if "MA" in annotation_var:
             annotation_name = "".join(
                 [
